Split lazy loading from functionality
Generated-by: claude-code Sonnet 4.6
